Add files via upload
1 equals paper, 2 equals stone, and 3 equals scissors. If it has any value other than one of these values, it will exit the game.
Thus, it points to a move, generate a random number between 1 and 3.

Based on the option generated by the computer and a user-informed option, the application may have been removed, the user was chosen, or the computer won.

After showing the result, ask for a new option (item 1). When the user chooses to exit the game, that is, when he enters anything other than 1, 2 or 3, show on the screen the times played against the computer.